+++ sdk/disconnectedoperations/arm-disconnectedoperations/src/restorePollerHelpers.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import type { EdgeClient } from "./edgeClient.js";
+import type { DisconnectedOperationsMgmtClient } from "./disconnectedOperationsMgmtClient.js";
 import {
   _$deleteDeserialize,
   _createOrUpdateDeserialize,
@@ -36,7 +36,7 @@ export interface RestorePollerOptions<
  * needs to be constructed after the original one is not in scope.
  */
 export function restorePoller<TResponse extends PathUncheckedResponse, TResult>(
-  client: EdgeClient,
+  client: DisconnectedOperationsMgmtClient,
   serializedState: string,
   sourceOperation: (...args: any[]) => PollerLike<OperationState<TResult>, TResult>,
   options?: RestorePollerOptions<TResult>,